refactor: split test in dedicated directory
